minor refactoring and minimal game test
