+++ python/tests/testware/selenium_smoke_test.py
@@ -1,22 +1,12 @@
 import unittest
 
-from os import environ
-from selenium.webdriver import Firefox
-from selenium.webdriver.firefox.options import Options as FirefoxOptions
-from selenium.webdriver.remote.webdriver import WebDriver
+from python.testware.seleniumwrapper import SeleniumBaseTest
 
 
-class SeleniumSmokeTest(unittest.TestCase):
-
-    def setUp(self):
-        options: FirefoxOptions = FirefoxOptions()
-        if "GITHUB_ACTIONS" in environ:
-            options.headless = True
-        self.browser: WebDriver = Firefox(options=options)
-        self.addCleanup(self.browser.quit)
+class SeleniumSmokeTest(SeleniumBaseTest):
 
     def test_can_run_firefox_and_quit(self):
-        pass
+        self.assertEqual("firefox", self.browser.name)
 
 
 if __name__ == '__main__':
